requirements.txt
bleach==6.1.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
contourpy==1.3.0
cycler==0.12.1
filelock==3.16.1
fonttools==4.54.1
fsspec==2024.9.0
idna==3.10
importlib_resources==6.4.5
Jinja2==3.1.4
kaggle==1.6.17
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.9.2
mpmath==1.3.0
networkx==3.2.1
numpy==2.0.2
opendatasets==0.1.22
packaging==24.1
pillow==11.0.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-slugify==8.0.4
PyYAML==6.0.2
requests==2.32.3
six==1.16.0
sympy==1.13.1
text-unidecode==1.3
torch==2.5.0
torchvision==0.20.0
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.3
webencodings==0.5.1
zipp==3.20.2
